Fully flesh out voting with historical records
The big key was actually hooking into the client reaction events
directly, and keeping records in the database. Since nothing is
preventing users from removing their vote, or voting multiple times,
that needed to be accounted for.

Things seem to be working pretty well, but the voting library will
likely need to be enhanced, adding a callback with poll results fired on
stop();